Optionally skip plv8 extension install
The newer postgis/postgres docker images do not appear to have plv8 as
an option. I cannot find a reference to use using plv8 (our sql
functions use plpgsql) so I added a precondition to check if we are able to
install, otherwise we continue. NOTE: all tests of the nldi services
against this new image still pass.